+++ src/DfE.ExternalApplications.Web/Pages/Admin/Admin.cshtml
@@ -1,5 +1,5 @@
-@page
-@model DfE.ExternalApplications.Web.Pages.AdminModel
+@page "/admin"
+@model DfE.ExternalApplications.Web.Pages.Admin.AdminModel
 @{
     ViewData["Title"] = "Admin - Template Management";
 }
