+++ src/DfE.ExternalApplications.Web/Pages/Admin/TemplateManager.cshtml
@@ -1,5 +1,5 @@
-@page
-@model DfE.ExternalApplications.Web.Pages.TemplateManagerModel
+@page "/admin/template-manager"
+@model DfE.ExternalApplications.Web.Pages.Admin.TemplateManagerModel
 @{
     ViewData["Title"] = "Template Manager";
 }
