+++ src/DfE.ExternalApplications.Web/Pages/Applications/Dashboard.cshtml
@@ -1,4 +1,4 @@
-﻿@page "/dashboard"
+﻿@page "/applications/dashboard"
 @using DfE.CoreLibs.Contracts.ExternalApplications.Enums
 @using DfE.CoreLibs.Contracts.ExternalApplications.Models.Response
 @using GovUk.Frontend.AspNetCore.TagHelpers
@@ -73,7 +73,7 @@
                                 </strong>
                             </td>
                             <td class="govuk-table__cell">
-                                <a href="/render-form/@app.ApplicationReference" class="govuk-link">@action</a>
+                                <a href="/form-engine/render-form/@app.ApplicationReference" class="govuk-link">@action</a>
                             </td>
                         </tr>
                     }
