+++ src/DfE.ExternalApplications.Web/Pages/Index.cshtml
@@ -252,7 +252,7 @@
               <li>the date you want to transfer</li>
             </ul>
 
-            <a href="/dashboard" role="button" draggable="false" class="govuk-button govuk-button--start" data-module="govuk-button">
+            <a href="/applications/dashboard" role="button" draggable="false" class="govuk-button govuk-button--start" data-module="govuk-button">
                 Start
                 <svg class="govuk-button__start-icon" xmlns="http://www.w3.org/2000/svg" width="17.5" height="19" viewBox="0 0 33 40" aria-hidden="true" focusable="false">
                     <path fill="currentColor" d="M0 0h13l20 20-20 20H0l20-20z" />
