+++ src/DfE.ExternalApplications.Web/Pages/Logout.cshtml
@@ -17,7 +17,7 @@
                 Sign out
             </button>
             
-            <a href="/Dashboard" class="govuk-link govuk-body">Cancel</a>
+            <a href="/applications/dashboard" class="govuk-link govuk-body">Cancel</a>
         </form>
 
         <details class="govuk-details" data-module="govuk-details">
