+++ src/DfE.ExternalApplications.Web/Pages/TestLogout.cshtml
@@ -25,7 +25,7 @@
                 Sign out
             </button>
             
-            <a href="/Dashboard" class="govuk-link govuk-body">Cancel</a>
+            <a href="/applications/dashboard" class="govuk-link govuk-body">Cancel</a>
         </form>
     </div>
 </div>
